\name{getPhenCenters}
\alias{getPhenCenters}
\title{Method "getPhenCenters"}
\description{
    Returns the list of IMPC phenotyping centers. 
}
\usage{
    getPhenCenters()
}

\value{
    Returns the list of IMPC phenotyping centers.
}

\author{Natalja Kurbatova, Jeremy Mason}
\example{
    getPhenCenters()
}