\name{printPipelines}
\alias{printPipelines}
\title{Method "printPipelines"}
\description{
    Prints out IMPC pipelines (IDS and names) that are run by specified phenotyping center.
}
\usage{
    printPipelines(PhenCenterName=NULL)
}
\arguments{
    \item{PhenCenterName}{IMPC phenotyping center; mandatory argument}
}

\author{Natalja Kurbatova, Jeremy Mason}
\example{
    printPipelines("WTSI")
}
